% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famTEMsel.main.R
\name{make_ITR_famTEMsel}
\alias{make_ITR_famTEMsel}
\title{make individualized treatment recommendations (ITRs) based on a \code{famTEMsel} object}
\usage{
make_ITR_famTEMsel(famTEMsel.obj, newX = NULL, newZ = NULL,
  lambda.index = NULL, maximize = TRUE)
}
\arguments{
\item{famTEMsel.obj}{a \code{famTEMsel} object, containing the fitted constrained functional additive models.}

\item{newX}{a length-p list of new values for the functional-valued covariates X, where the jth element is a (n-by-n.eval[j]) matrix of the observed jth function, at which predictions are to be made; if \code{NULL}, X from the training set is used.}

\item{newZ}{a (n-by-q) matrix of new values for the scalar-valued covariates Z at which predictions are to be made; if \code{NULL}, Z from the training set is used.}

\item{lambda.index}{an index of the regularization parameter \code{lambda} at which predictions are to be made; one can supply \code{lambda.opt.index} obtained from the function \code{cv.famTEMsel()}; the default is \code{NULL}, in which case the predictions are made at the \code{lambda.index} used in obtaining \code{famTEMsel.obj}.}

\item{maximize}{default is \code{TRUE}, assuming a larger value of the outcome is better; if \code{FALSE}, a smaller value is assumed to be prefered.}
}
\value{
\item{pred.new}{a (n-by-L) matrix of predicted values, with each column representing one of the L treatment options.}
\item{trt.rule}{a (n-by-1) vector of the individualized treatment recommendations}
}
\description{
The function \code{make_ITR_famTEMsel} returns individualized treatment decision recommendations for subjects with pretreatment characteristics \code{newX} and \code{newZ}, given a \code{famTEMsel} object, \code{famTEMsel.obj}, and an (optimal) regularization parameter index, \code{lambda.index}.
}
\seealso{
\code{famTEMsel},\code{cv.famTEMsel}, \code{predict_famTEMsel}
}
\author{
Park, Petkova, Tarpey, Ogden
}
