% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famTEMsel.main.R
\name{plot_famTEMsel}
\alias{plot_famTEMsel}
\title{plot component functions from a \code{famTEMsel} object}
\usage{
plot_famTEMsel(famTEMsel.obj, newX = NULL, newZ = NULL, newA = NULL,
  scatter.plot = TRUE, lambda.index = famTEMsel.obj$lambda.index,
  which.index = famTEMsel.obj$nonzero.index, ylims,
  single.index.plot = TRUE, solution.path = FALSE)
}
\arguments{
\item{famTEMsel.obj}{a \code{famTEMsel} object}

\item{newX}{a (n by p) list of new values for the functional covariates X at which predictions are to be made; the jth element of the list corresponds to a n-by-n.eval[j] matrix of the observed jth functional covariates; n.eval[j] represents the number of evaluation points of the jth functional covariates; if \code{NULL}, X from the training set is used.}

\item{newZ}{a (n by q) matrix of new values for the scalar covariates Z at which predictions are to be made; if \code{NULL}, Z from the training set is used.}

\item{newA}{a (n-by-1) vector of new values for the treatment A at which plots are to be made; the default is \code{NULL}, in which case A is taken from the training set.}

\item{scatter.plot}{if \code{TRUE}, draw scatter plots of partial residuals versus the covariates; these scatter plots are made based on the training observations;  the default is \code{TRUE}.}

\item{lambda.index}{an index of the tuning parameter \code{lambda} at which plots are to be made; one can supply \code{lambda.opt.index} obtained from the function \code{cv.samTEMsel}; the default is \code{NULL}, in which case \code{plot_samTEMsel} utilizes the most non-sparse model.}

\item{which.index}{this specifies which component functions are to be plotted; the default is all p component functions, i.e., 1:p.}

\item{ylims}{this specifies the vertical range of the plots, e.g., c(-10, 10).}

\item{single.index.plot}{if \code{TRUE}, draw the plots of the estimated single-index coefficient functions; the default is \code{TRUE}.}

\item{solution.path}{if \code{TRUE}, draw the functional norms of the fitted component functions (based on the training set) versus the regularization parameter; the default is \code{FALSE}.}
}
\description{
Produces plots of the component functions and the single-index coefficient functions from a \code{famTEMsel} object.
}
\seealso{
\code{famTEMsel},\code{predict_famTEMsel}, \code{cv.famTEMsel}
}
\author{
Park, Petkova, Tarpey, Ogden
}
