% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famTEMsel.main.R
\name{predict_famTEMsel}
\alias{predict_famTEMsel}
\title{\code{famTEMsel} prediction function}
\usage{
predict_famTEMsel(famTEMsel.obj, newX = NULL, newZ = NULL,
  newA = NULL, type = "response", lambda.index = NULL)
}
\arguments{
\item{famTEMsel.obj}{a \code{famTEMsel} object}

\item{newX}{a (n by p) list of new values for the functional covariates X at which predictions are to be made; the jth element of the list corresponds to a n-by-n.eval[j] matrix of the observed jth functional covariates; n.eval[j] represents the number of evaluation points of the jth functional covariates; if \code{NULL}, X from the training set is used.}

\item{newZ}{a (n by q) matrix of new values for the scalar covariates Z at which predictions are to be made; if \code{NULL}, Z from the training set is used.}

\item{newA}{a (n by 1) vector of new values for the treatment A at which predictions are to be made; if \code{NULL}, A from the training set is used.}

\item{type}{the type of prediction required; the default "response" gives the predicted responses y based on the whole model; the alternative "terms" gives the component-wise predicted responses from each of the p components (and plus the treatment-specific intercepts) of the model.}

\item{lambda.index}{an index of the tuning parameter \code{lambda} at which predictions are to be made; one can supply \code{lambda.opt.index} obtained from the function \code{cv.samTEMsel}; the default is \code{NULL}, in which case the predictions based on the most non-sparse model is returned.}
}
\value{
\item{predicted}{a (n-by-\code{length(lambda.index)}) matrix of predicted values; a (n-by-\code{length(lambda.index)}*(p+q+1)) matrix of predicted values if type =  "terms", where the last column corresponds to the (treatment-specific) intercept.}
\item{U}{a n-by-(p+q) matrix of the index variables; the first p columns correspond to the 1-D projections of the p functional covariates and the last q columns correspond to the q scalar covariates.}
}
\description{
\code{predict_famTEMsel} makes predictions given a (new) set of functional covariates \code{newX}, a (new) set of scalar covariates \code{newZ} and a (new) vector of treatment indicators \code{newA} based on a constrained functional additive model \code{famTEMsel.obj}.
Specifically, \code{predict_famTEMsel} predicts the responses y based on the (X,Z)-by-A interaction effect (plus the A main effect) portion of the full model that includes the unspecified X main effect term.
}
\seealso{
\code{famTEMsel},\code{cv.famTEMsel}, \code{plot_famTEMsel}
}
\author{
Park, Petkova, Tarpey, Ogden
}
